pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dalaipy"
version = "2.0.1"
authors = [
  { name="William Stella", email="william.a.stella@gmail.com" },
]
description = "A Python Wrapper for Dalai"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["python-socketio"]

[project.urls]
"Homepage" = "https://github.com/wastella/dalaipy"
"Bug Tracker" = "https://github.com/wastella/dalaipy/issues"
